+++ src/main/java/frc/states/MechState.java
@@ -1,20 +1,29 @@
 package frc.states;
 
-public abstract class MechState {
-    public void setFlywheelOff() {
+import frc.robot.constants.LauncherConstants.LauncherMode;
+import frc.robot.subsystems.LauncherSubsystem;
 
+public abstract class MechState {
+    protected final LauncherSubsystem mLauncherSubsystem;
+
+    public MechState (LauncherSubsystem launcherSubsystem) {
+        mLauncherSubsystem = launcherSubsystem;
+    }
+
+    public void setFlywheelOff() {
+        mLauncherSubsystem.stopLauncher();
     }
 
     public void setSpeakerSP() {
-
+        mLauncherSubsystem.setLauncherMode(LauncherMode.SPEAKER);
     }
 
     public void setAmpSP() {
-
+        mLauncherSubsystem.setLauncherMode(LauncherMode.AMP);
     }
 
     public void setTrapSP() {
-
+        mLauncherSubsystem.setLauncherMode(LauncherMode.TRAP);
     }
 
     public void intake() {
