+++ src/main/java/frc/states/MechStates/ReadyForIntake.java
@@ -1,5 +1,5 @@
 package frc.states.MechStates;
-import frc.robot.constants.LauncherConstants.LauncherMode;
+
 import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.subsystems.LauncherSubsystem;
 import frc.states.MechState;
@@ -7,34 +7,12 @@ import frc.states.MechState;
 public class ReadyForIntake extends MechState {
 
     private final IntakeSubsystem mIntakeSubsystem;
-    private final LauncherSubsystem mLauncherSubsystem;
-
 
     public ReadyForIntake (IntakeSubsystem intakeSubsystem, LauncherSubsystem launcherSubsystem) {
+        super(launcherSubsystem);
         mIntakeSubsystem = intakeSubsystem;
-        mLauncherSubsystem = launcherSubsystem;
     }
 
-    @Override
-    public void setFlywheelOff() {
-        mLauncherSubsystem.stopLauncher();
-    }
-
-    @Override
-    public void setSpeakerSP() {
-        mLauncherSubsystem.setLauncherMode(LauncherMode.SPEAKER);
-    }
-
-    @Override
-    public void setAmpSP() {
-        mLauncherSubsystem.setLauncherMode(LauncherMode.AMP);
-    }
-
-    @Override
-    public void setTrapSP() {
-        mLauncherSubsystem.setLauncherMode(LauncherMode.TRAP);
-    }
-    
     @Override
     public void intake() {
         mIntakeSubsystem.getAutomousIntakeCommand();
